+++ packages/cozy-harvest-lib/src/components/TriggerLauncher.jsx
@@ -1,15 +1,14 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 
-import { withMutations } from 'cozy-client'
+import { withClient, withMutations } from 'cozy-client'
 
 import TwoFAModal from './TwoFAModal'
 import { accountsMutations } from '../connections/accounts'
 import { triggersMutations } from '../connections/triggers'
-import withKonnectorJob from './hoc/withKonnectorJob'
 import withLocales from './hoc/withLocales'
 
-import {
+import KonnectorJob, {
   ERROR_EVENT,
   SUCCESS_EVENT,
   TWO_FA_REQUEST_EVENT,
@@ -39,20 +38,29 @@ export class TriggerLauncher extends Component {
     super(props, context)
     this.state = { showTwoFAModal: false }
 
-    this.dismissTwoFAModal = this.dismissTwoFAModal.bind(this)
-    this.displayTwoFAModal = this.displayTwoFAModal.bind(this)
+    this.handleTwoFA = this.handleTwoFA.bind(this)
+    this.handleError = this.handleError.bind(this)
+    this.handleSuccess = this.handleSuccess.bind(this)
+
+    this.launch = this.launch.bind(this)
   }
 
-  componentDidMount() {
-    this.props.konnectorJob
-      .on(ERROR_EVENT, this.dismissTwoFAModal)
-      .on(SUCCESS_EVENT, this.dismissTwoFAModal)
-      .on(TWO_FA_REQUEST_EVENT, this.displayTwoFAModal)
-      .on(TWO_FA_MISMATCH_EVENT, this.displayTwoFAModal)
-  }
+  launch() {
+    const { client, trigger } = this.props
+    const konnectorJob = new KonnectorJob(client, trigger)
+    konnectorJob
+      .on(ERROR_EVENT, this.handleError)
+      .on(SUCCESS_EVENT, this.handleSuccess)
+      .on(TWO_FA_REQUEST_EVENT, this.handleTwoFA)
+      .on(TWO_FA_MISMATCH_EVENT, this.handleTwoFA)
 
-  componentWillUnmount() {
-    this.props.konnectorJob.unwatch()
+    this.setState({
+      error: null,
+      konnectorJob,
+      running: true
+    })
+
+    konnectorJob.launch()
   }
 
   dismissTwoFAModal() {
@@ -64,15 +72,36 @@ export class TriggerLauncher extends Component {
     this.setState({ showTwoFAModal: true })
   }
 
-  render() {
-    const { showTwoFAModal } = this.state
-    const { children, konnectorJob, submitting } = this.props
+  handleError(error) {
+    this.dismissTwoFAModal()
+    this.setState({ error, running: false })
+    this.stopWatchingKonnectorJob()
+  }
 
+  handleSuccess() {
+    this.dismissTwoFAModal()
+    this.setState({ running: false })
+    this.stopWatchingKonnectorJob()
+  }
+
+  handleTwoFA() {
+    this.displayTwoFAModal()
+  }
+
+  stopWatchingKonnectorJob() {
+    const { konnectorJob } = this.state
+    konnectorJob.unwatch()
+  }
+
+  render() {
+    const { error, running, showTwoFAModal } = this.state
+    const { children, konnectorJob, submitting } = this.props
     return (
       <div>
         {children({
-          launch: konnectorJob.launch,
-          running: konnectorJob.isRunning() || submitting
+          error,
+          launch: this.launch,
+          running: !!running || !!submitting
         })}
         {showTwoFAModal && (
           <TwoFAModal
@@ -92,9 +121,9 @@ TriggerLauncher.propTypes = {
    */
   children: PropTypes.func.isRequired,
   /**
-   * The konnectorJob instance provided by withKonnectorJob
+   * CozyClient instance
    */
-  konnectorJob: PropTypes.object.isRequired,
+  client: PropTypes.object.isRequired,
   /**
    * Indicates if trigger is already runnning
    */
@@ -102,7 +131,7 @@ TriggerLauncher.propTypes = {
 }
 
 export default withLocales(
-  withKonnectorJob(
+  withClient(
     withMutations(accountsMutations, triggersMutations)(TriggerLauncher)
   )
 )