+++ packages/cozy-procedures/src/index.js
@@ -1,16 +1,20 @@
 import React from 'react'
 import { Route, IndexRoute } from 'react-router'
 
-import Procedure from './Procedure'
+import App from './App'
 import Overview from './containers/Overview'
+import Amount from './components/Amount'
+import Duration from './components/Duration'
 import PersonalDataForm from './containers/PersonalDataForm'
-import OtherForm from './components/OtherComp'
+import Documents from './components/Documents'
 
 const injectProcedureRoutes = ({ root }) => (
-  <Route path={root} component={Procedure}>
+  <Route path={root} component={App}>
     <IndexRoute component={() => <Overview />} />
+    <Route path="amount" component={() => <Amount />} />
+    <Route path="duration" component={() => <Duration />} />
     <Route path="personal" component={() => <PersonalDataForm />} />
-    <Route path="other" component={OtherForm} />
+    <Route path="documents" component={() => <Documents />} />
   </Route>
 )