+++ packages/cozy-harvest-lib/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.59.0](https://github.com/cozy/cozy-libs/compare/cozy-harvest-lib@0.58.0...cozy-harvest-lib@0.59.0) (2019-06-20)
+
+
+### Features
+
+* **harvest:** Fetch job right after subscribing realtime 📝 ([25b894d](https://github.com/cozy/cozy-libs/commit/25b894d))
+
+
+
+
+
+# [0.58.0](https://github.com/cozy/cozy-libs/compare/cozy-harvest-lib@0.57.3...cozy-harvest-lib@0.58.0) (2019-06-20)
+
+
+### Features
+
+* **harvest:** Add LaunchTriggerCard ✨ ([d9fdf4f](https://github.com/cozy/cozy-libs/commit/d9fdf4f))
+
+
+
+
+
+## [0.57.3](https://github.com/cozy/cozy-libs/compare/cozy-harvest-lib@0.57.2...cozy-harvest-lib@0.57.3) (2019-06-20)
+
+
+### Bug Fixes
+
+* **harvest:** Make TriggerLauncher detects success again 🚑 ([d2487ee](https://github.com/cozy/cozy-libs/commit/d2487ee))
+
+
+
+
+
+## [0.57.2](https://github.com/cozy/cozy-libs/compare/cozy-harvest-lib@0.57.1...cozy-harvest-lib@0.57.2) (2019-06-20)
+
+
+### Bug Fixes
+
+* :pencil2: Missing spread ([7947f43](https://github.com/cozy/cozy-libs/commit/7947f43))
+
+
+
+
+
+## [0.57.1](https://github.com/cozy/cozy-libs/compare/cozy-harvest-lib@0.57.0...cozy-harvest-lib@0.57.1) (2019-06-20)
+
+
+### Bug Fixes
+
+* **harvest:** :bug: Omit also i18n props from withLocales ([5f88752](https://github.com/cozy/cozy-libs/commit/5f88752))
+
+
+
+
+
 # [0.57.0](https://github.com/cozy/cozy-libs/compare/cozy-harvest-lib@0.56.0...cozy-harvest-lib@0.57.0) (2019-06-18)